Se relacionaron los Entitys aunque no se cambió la estructura de rol y permisos

+++ ProyectoMDW_Completo/demo/src/main/java/com/marcosdeDesarrollo/Ropa/Domain/DTO/CategoriaDto.java
@@ -4,16 +4,16 @@ import com.marcosdeDesarrollo.Ropa.Persistencia.Entity.Estado;
 import java.time.LocalDateTime;
 
 public class CategoriaDto {
-    private Long idCategoria;
+    private Integer idCategoria;
     private String nombreCategoria;
     private String descripcion;
     private Estado estado;
     private LocalDateTime fechaCreacion;
 
-    public Long getIdCategoria() {
+    public Integer getIdCategoria() {
         return idCategoria;
     }
-    public void setIdCategoria(Long idCategoria) {
+    public void setIdCategoria(Integer idCategoria) {
         this.idCategoria = idCategoria;
     }
